+++ govtool/frontend/src/context/featureFlag.tsx
@@ -1,7 +1,38 @@
-import { PropsWithChildren, useMemo, createContext, useContext } from "react";
+import {
+  PropsWithChildren,
+  useMemo,
+  createContext,
+  useContext,
+  useCallback,
+} from "react";
 
-const FeatureFlagContext = createContext({
+import { GovernanceActionType } from "@/types/governanceAction";
+
+import { useAppContext } from "./appContext";
+
+const govActionVotingEnabledSinceProtocolVersion = {
+  [GovernanceActionType.InfoAction]: 9,
+  // TODO: Add minimum protocol versions for the following actions
+  [GovernanceActionType.HardForkInitiation]: Number.MAX_SAFE_INTEGER,
+  [GovernanceActionType.NewCommittee]: Number.MAX_SAFE_INTEGER,
+  [GovernanceActionType.NewConstitution]: Number.MAX_SAFE_INTEGER,
+  [GovernanceActionType.NoConfidence]: Number.MAX_SAFE_INTEGER,
+  [GovernanceActionType.ParameterChange]: Number.MAX_SAFE_INTEGER,
+  [GovernanceActionType.TreasuryWithdrawals]: Number.MAX_SAFE_INTEGER,
+};
+/**
+ * The feature flag context type.
+ */
+type FeatureFlagContextType = {
+  isProposalDiscussionForumEnabled: boolean;
+  isVotingOnGovernanceActionEnabled: (
+    governanceActionType: GovernanceActionType,
+  ) => boolean;
+};
+
+const FeatureFlagContext = createContext<FeatureFlagContextType>({
   isProposalDiscussionForumEnabled: false,
+  isVotingOnGovernanceActionEnabled: () => false,
 });
 
 /**
@@ -10,13 +41,28 @@ const FeatureFlagContext = createContext({
  * @param children - The child components to render.
  */
 const FeatureFlagProvider = ({ children }: PropsWithChildren) => {
+  const { epochParams, isAppInitializing } = useAppContext();
+
+  /**
+   * Determines if voting on a governance action is enabled based on the protocol version.
+   * @param governanceActionType - The type of governance action.
+   * @returns A boolean indicating whether voting is enabled for the specified governance action.
+   */
+  const isVotingOnGovernanceActionEnabled = useCallback(
+    (governanceActionType: GovernanceActionType) =>
+      (epochParams?.protocol_major || 0) >
+      govActionVotingEnabledSinceProtocolVersion[governanceActionType],
+    [isAppInitializing],
+  );
+
   const value = useMemo(
     () => ({
       isProposalDiscussionForumEnabled:
         import.meta.env.VITE_IS_PROPOSAL_DISCUSSION_FORUM_ENABLED === "true" ||
         false,
+      isVotingOnGovernanceActionEnabled,
     }),
-    [],
+    [isVotingOnGovernanceActionEnabled],
   );
 
   return (