+++ govtool/frontend/src/services/requests/getEpochParams.ts
@@ -1,7 +1,9 @@
+import { EpochParams } from "@/models";
+
 import { API } from "../API";
 
 export const getEpochParams = async () => {
-  const response = await API.get("/epoch/params");
+  const response = await API.get<EpochParams>("/epoch/params");
 
   return response.data;
 };